requirements.txt
cfgv==3.4.0
distlib==0.3.9
exceptiongroup==1.2.2
filelock==3.18.0
identify==2.6.10
iniconfig==2.1.0
narwhals==1.36.0
nodeenv==1.9.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
pre-commit==4.2.0
psutil==7.0.0
psycopg2-binary==2.9.10
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
ruff==0.11.7
six==1.17.0
tomli==2.2.1
tzdata==2025.2
virtualenv==20.30.0
